detail view implemented time log
